modificado:     ../api/src/control/index.js 	modificado:     ../api/src/models/Country.js 	modificado:     ../api/src/routes/activities.js

+++ api/src/models/Country.js
@@ -56,7 +56,7 @@ module.exports = (sequelize) => {
       type: DataTypes.STRING
     },
     area:{
-      type: DataTypes.BIGINT
+      type: DataTypes.FLOAT
     },
     pob:{
       type: DataTypes.INTEGER
